+++ rebootpy/enums.py
@@ -185,7 +185,7 @@ class UserSearchPlatform(Enum):
 
 
 class UserSearchMatchType(Enum):
-    EXACT = 'exact'
+    EXACT  = 'exact'
     PREFIX = 'prefix'
 
 
@@ -203,131 +203,133 @@ class AwayStatus(Enum):
 
 
 class SeasonStartTimestamp(Enum):
-    SEASON_1  = 1508889601
-    SEASON_2  = 1513209601
-    SEASON_3  = 1519257601
-    SEASON_4  = 1525132801
-    SEASON_5  = 1531353601
-    SEASON_6  = 1538006401
-    SEASON_7  = 1544054401
-    SEASON_8  = 1551312001
-    SEASON_9  = 1557360001
-    SEASON_10 = 1564617601
-    SEASON_11 = 1571097601
-    SEASON_12 = 1582156801
-    SEASON_13 = 1592352001
-    SEASON_14 = 1598486401
-    SEASON_15 = 1606867201
-    SEASON_16 = 1615852801
-    SEASON_17 = 1623110401
-    SEASON_18 = 1631491201
-    SEASON_19 = 1638716400
-    SEASON_20 = 1647763140
-    SEASON_21 = 1654430400
-    SEASON_22 = 1663503780
-    SEASON_23 = 1670157000
-    SEASON_24 = 1678453200
-    SEASON_25 = 1686313800
-    SEASON_26 = 1692966600
-    SEASON_27 = 1699016400
-    SEASON_28 = 1701577740
-    SEASON_29 = 1709951400
-    SEASON_30 = 1716508800
+    C1S1   = 1508889601
+    C1S2   = 1513209601
+    C1S3   = 1519257601
+    C1S4   = 1525132801
+    C1S5   = 1531353601
+    C1S6   = 1538006401
+    C1S7   = 1544054401
+    C1S8   = 1551312001
+    C1S9   = 1557360001
+    C1SX   = 1564617601
+    C2S1   = 1571097601
+    C2S2   = 1582156801
+    C2S3   = 1592352001
+    C2S4   = 1598486401
+    C2S5   = 1606867201
+    C2S6   = 1615852801
+    C2S7   = 1623110401
+    C2S8   = 1631491201
+    C3S1   = 1638716400
+    C3S2   = 1647763140
+    C3S3   = 1654430400
+    C3S4   = 1663503780
+    C4S1   = 1670157000
+    C4S2   = 1678453200
+    C4S3   = 1686313800
+    C4S4   = 1692966600
+    C4SOG  = 1699016400
+    C5S1   = 1701577740
+    C5S2   = 1709951400
+    C5S3   = 1716508800
+    C5S4   = 1723802400
 
 
 class SeasonEndTimestamp(Enum):
-    SEASON_1  = 1513123200
-    SEASON_2  = 1519171200
-    SEASON_3  = 1525046400
-    SEASON_4  = 1531353600
-    SEASON_5  = 1538006400
-    SEASON_6  = 1544054400
-    SEASON_7  = 1551312000
-    SEASON_8  = 1557360000
-    SEASON_9  = 1564617600
-    SEASON_10 = 1570924800
-    SEASON_11 = 1582156800
-    SEASON_12 = 1592352000
-    SEASON_13 = 1598486400
-    SEASON_14 = 1606867200
-    SEASON_15 = 1615852800
-    SEASON_16 = 1623110400
-    SEASON_17 = 1631491200
-    SEASON_18 = 1638651540
-    SEASON_19 = 1647759540
-    SEASON_20 = 1654372800
-    SEASON_21 = 1663479000
-    SEASON_22 = 1670140800
-    SEASON_23 = 1678429800
-    SEASON_24 = 1686288600
-    SEASON_25 = 1692945000
-    SEASON_26 = 1698994800
-    SEASON_27 = 1701576000
-    SEASON_28 = 1709794800
-    SEASON_29 = 1716533100
+    C1S1  = 1513123200
+    C1S2  = 1519171200
+    C1S3  = 1525046400
+    C1S4  = 1531353600
+    C1S5  = 1538006400
+    C1S6  = 1544054400
+    C1S7  = 1551312000
+    C1S8  = 1557360000
+    C1S9  = 1564617600
+    C1SX  = 1570924800
+    C2S1  = 1582156800
+    C2S2  = 1592352000
+    C2S3  = 1598486400
+    C2S4  = 1606867200
+    C2S5  = 1615852800
+    C2S6  = 1623110400
+    C2S7  = 1631491200
+    C2S8  = 1638651540
+    C3S1  = 1647759540
+    C3S2  = 1654372800
+    C3S3  = 1663479000
+    C3S4  = 1670140800
+    C4S1  = 1678429800
+    C4S2  = 1686288600
+    C4S3  = 1692945000
+    C4S4  = 1698994800
+    C4SOG = 1701576000
+    C5S1  = 1709794800
+    C5S2  = 1716533100
+    C5S3  = 1723777200
 
 
 class BattlePassStat(Enum):
-    SEASON_11 = ('s11_social_bp_level', SeasonEndTimestamp.SEASON_11.value)
-    SEASON_12 = ('s11_social_bp_level', SeasonEndTimestamp.SEASON_12.value)
-    SEASON_13 = (('s13_social_bp_level', 's11_social_bp_level'), SeasonEndTimestamp.SEASON_13.value)
-    SEASON_14 = ('s14_social_bp_level', SeasonEndTimestamp.SEASON_14.value)
-    SEASON_15 = ('s15_social_bp_level', SeasonEndTimestamp.SEASON_15.value)
-    SEASON_16 = ('s16_social_bp_level', SeasonEndTimestamp.SEASON_16.value)
-    SEASON_17 = ('s17_social_bp_level', SeasonEndTimestamp.SEASON_17.value)
-    SEASON_18 = ('s18_social_bp_level', SeasonEndTimestamp.SEASON_18.value)
-    SEASON_19 = ('s19_social_bp_level', SeasonEndTimestamp.SEASON_19.value)
-    SEASON_20 = ('s20_social_bp_level', SeasonEndTimestamp.SEASON_20.value)
-    SEASON_21 = ('s21_social_bp_level', SeasonEndTimestamp.SEASON_21.value)
-    SEASON_22 = ('s22_social_bp_level', SeasonEndTimestamp.SEASON_22.value)
-    SEASON_23 = ('s23_social_bp_level', SeasonEndTimestamp.SEASON_23.value)
-    SEASON_24 = ('s24_social_bp_level', SeasonEndTimestamp.SEASON_24.value)
-    SEASON_25 = ('s25_social_bp_level', SeasonEndTimestamp.SEASON_25.value)
-    SEASON_26 = ('s26_social_bp_level', SeasonEndTimestamp.SEASON_26.value)
-    SEASON_27 = ('s27_social_bp_level', SeasonEndTimestamp.SEASON_27.value)
-    SEASON_28 = ('s28_social_bp_level', SeasonEndTimestamp.SEASON_28.value)
-    SEASON_29 = ('s29_social_bp_level', SeasonEndTimestamp.SEASON_29.value)
+    C2S1  = ('s11_social_bp_level', SeasonEndTimestamp.C2S1.value)
+    C2S2  = ('s11_social_bp_level', SeasonEndTimestamp.C2S2.value)
+    C2S3  = (('s13_social_bp_level', 's11_social_bp_level'), SeasonEndTimestamp.C2S3.value)
+    C2S4  = ('s14_social_bp_level', SeasonEndTimestamp.C2S4.value)
+    C2S5  = ('s15_social_bp_level', SeasonEndTimestamp.C2S5.value)
+    C2S6  = ('s16_social_bp_level', SeasonEndTimestamp.C2S6.value)
+    C2S7  = ('s17_social_bp_level', SeasonEndTimestamp.C2S7.value)
+    C2S8  = ('s18_social_bp_level', SeasonEndTimestamp.C2S8.value)
+    C3S1  = ('s19_social_bp_level', SeasonEndTimestamp.C3S1.value)
+    C3S2  = ('s20_social_bp_level', SeasonEndTimestamp.C3S2.value)
+    C3S3  = ('s21_social_bp_level', SeasonEndTimestamp.C3S3.value)
+    C3S4  = ('s22_social_bp_level', SeasonEndTimestamp.C3S4.value)
+    C4S1  = ('s23_social_bp_level', SeasonEndTimestamp.C4S1.value)
+    C4S2  = ('s24_social_bp_level', SeasonEndTimestamp.C4S2.value)
+    C4S3  = ('s25_social_bp_level', SeasonEndTimestamp.C4S3.value)
+    C4S4  = ('s26_social_bp_level', SeasonEndTimestamp.C4S4.value)
+    C4SOG = ('s27_social_bp_level', SeasonEndTimestamp.C4SOG.value)
+    C5S1  = ('s28_social_bp_level', SeasonEndTimestamp.C5S1.value)
+    C5S2  = ('s29_social_bp_level', SeasonEndTimestamp.C5S2.value)
 
 
 class Seasons(Enum):
-    C4S2 = ('2776dc', '9d7ebd')
-    C4S3 = ('ggOwuK', 'AjRdrb')
-    C4S4 = ('gXffl', 'yHNFu')
+    C4S2  = ('2776dc', '9d7ebd')
+    C4S3  = ('ggOwuK', 'AjRdrb')
+    C4S4  = ('gXffl', 'yHNFu')
     C4SOG = ('OiK9k9', 'hEKWqj')
-    C5S1 = ('EYpme7', 'd0zEcd', 'dmd372')
-    C5S2 = ('ch3353', 'a1m0n3', 'rrwpwg')
-    C5S3 = ('N4PK1N', 'L1GHT5', 'rrzuel')
-    C5S4 = ('S4LT3D', 'P0T4T0', 'rr9qlw', 'M4rC4S', 'L4nC3r')
+    C5S1  = ('EYpme7', 'd0zEcd', 'dmd372')
+    C5S2  = ('ch3353', 'a1m0n3', 'rrwpwg')
+    C5S3  = ('N4PK1N', 'L1GHT5', 'rrzuel')
+    C5S4  = ('S4LT3D', 'P0T4T0', 'rr9qlw', 'M4rC4S', 'L4nC3r')
 
 
 class RankingType(Enum):
-    BATTLE_ROYALE = 'ranked-br'
-    ZERO_BUILD = 'ranked-zb'
-    ROCKET_RACING = 'delmar-competitive'
-    RELOAD = 'ranked_blastberry_build'
-    RELOAD_ZB = 'ranked_blastberry_nobuild'
+    BATTLE_ROYALE   = 'ranked-br'
+    ZERO_BUILD      = 'ranked-zb'
+    ROCKET_RACING   = 'delmar-competitive'
+    RELOAD          = 'ranked_blastberry_build'
+    RELOAD_ZB       = 'ranked_blastberry_nobuild'
 
 
 class Rank(Enum):
-    UNRANKED = None
-    BRONZE_1 = 0
-    BRONZE_2 = 1
-    BRONZE_3 = 2
-    SILVER_1 = 3
-    SILVER_2 = 4
-    SILVER_3 = 5
-    GOLD_1 = 6
-    GOLD_2 = 7
-    GOLD_3 = 8
-    PLATINUM_1 = 9
-    PLATINUM_2 = 10
-    PLATINUM_3 = 11
-    DIAMOND_1 = 12
-    DIAMOND_2 = 13
-    DIAMOND_3 = 14
-    ELITE = 15
-    CHAMPION = 16
-    UNREAL = 17
+    UNRANKED    = None
+    BRONZE_1    = 0
+    BRONZE_2    = 1
+    BRONZE_3    = 2
+    SILVER_1    = 3
+    SILVER_2    = 4
+    SILVER_3    = 5
+    GOLD_1      = 6
+    GOLD_2      = 7
+    GOLD_3      = 8
+    PLATINUM_1  = 9
+    PLATINUM_2  = 10
+    PLATINUM_3  = 11
+    DIAMOND_1   = 12
+    DIAMOND_2   = 13
+    DIAMOND_3   = 14
+    ELITE       = 15
+    CHAMPION    = 16
+    UNREAL      = 17
 
 
 class StatsCollectionType(Enum):